requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
constantly==23.10.4
cryptography==46.0.2
cssselect==1.3.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
fastapi==0.118.0
filelock==3.19.1
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.2
itemloaders==1.3.2
jiter==0.11.0
jmespath==1.0.1
lxml==6.0.2
motor==3.7.1
numpy==2.3.3
openai==2.1.0
packaging==25.0
parsel==1.10.0
playwright==1.55.0
portalocker==3.2.0
Protego==0.5.0
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==13.0.0
pymongo==4.15.2
pyOpenSSL==25.3.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
queuelib==1.8.0
requests==2.32.5
requests-file==2.1.0
Scrapy==2.13.3
scrapy-playwright==0.0.44
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
tldextract==5.3.0
tqdm==4.67.1
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.1.0
websockets==15.0.1
zope.interface==8.0.1
